extract interfaces from services and refactor
