+++ backend/src/main/java/pl/edu/agh/backend/compiler/files/RustConfigFile.java
@@ -2,7 +2,6 @@ package pl.edu.agh.backend.compiler.files;
 
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
-import pl.edu.agh.backend.compiler.files.RustFile;
 
 import java.io.File;
 import java.nio.file.Paths;
@@ -11,7 +10,9 @@ import java.util.List;
 @Getter
 @SuperBuilder
 public class RustConfigFile extends RustFile {
+
     protected String configFileName;
+
     protected String configContent;
 
     @Override
@@ -23,4 +24,5 @@ public class RustConfigFile extends RustFile {
     public String getPath() {
         return Paths.get(this.directory) + File.separator + configFileName;
     }
+
 }
