+++ backend/src/main/java/pl/edu/agh/backend/services/CompilerService.java
@@ -1,20 +1,11 @@
 package pl.edu.agh.backend.services;
 
-import org.springframework.stereotype.Service;
+import pl.edu.agh.backend.compiler.CompilationResponse;
 import pl.edu.agh.backend.compiler.files.RustConfigFile;
 import pl.edu.agh.backend.compiler.files.RustFile;
-import pl.edu.agh.backend.compiler.CompilationResponse;
-import pl.edu.agh.backend.compiler.ProgramProcess;
 
+public interface CompilerService {
+    CompilationResponse run(RustFile rustFile);
 
-@Service
-public class CompilerService {
-
-    public CompilationResponse run(RustFile rustFile) {
-        return new ProgramProcess(rustFile).runProcess();
-    }
-
-    public CompilationResponse configure(RustConfigFile rustFile) {
-        return new ProgramProcess(rustFile).buildConfig();
-    }
+    CompilationResponse configure(RustConfigFile rustFile);
 }
